Replace double usage with BigDecimal, to bring Vault2 further into the future.
Major clean up of javadocs in the Economy class.

+++ src/main/java/net/milkbowl/vault2/economy/EconomyResponse.java
@@ -15,10 +15,13 @@
  */
 package net.milkbowl.vault2.economy;
 
+import java.math.BigDecimal;
+
 /**
- * Indicates a typical Return for an Economy method.  
- * It includes a {@link ResponseType} indicating whether the plugin currently being used for Economy actually allows
- * the method, or if the operation was a success or failure.
+ * Indicates a typical Return for an Economy method. It includes a
+ * {@link ResponseType} indicating whether the plugin currently being used for
+ * Economy actually allows the method, or if the operation was a success or
+ * failure.
  *
  */
 public class EconomyResponse {
@@ -45,11 +48,11 @@ public class EconomyResponse {
     /**
      * Amount modified by calling method
      */
-    public final double amount;
+    public final BigDecimal amount;
     /**
      * New balance of account
      */
-    public final double balance;
+    public final BigDecimal balance;
     /**
      * Success or failure of call. Using Enum of ResponseType to determine valid
      * outcomes
@@ -67,7 +70,7 @@ public class EconomyResponse {
      * @param type Success or failure type of the operation
      * @param errorMessage Error message if necessary (commonly null)
      */
-    public EconomyResponse(double amount, double balance, ResponseType type, String errorMessage) {
+    public EconomyResponse(BigDecimal amount, BigDecimal balance, ResponseType type, String errorMessage) {
         this.amount = amount;
         this.balance = balance;
         this.type = type;
